requirements.txt
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg
et-xmlfile==1.1.0
filelock==3.13.1
gunicorn==22.0.0
platformdirs==4.2.0
prompt_toolkit==3.0.45
psycopg2==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-dateutil==2.9.0.post0
pytz==2022.7
redis==5.0.7
requests
sqlparse==0.4.4
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing_extensions==4.6.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.3
wcwidth==0.2.13
wheel==0.43.0
wrapt==1.15.0

nltk
Pillow==10.0.0
Pandas
numpy
geopy
openpyxl
scikit-learn

django-db-logger
elasticsearch-dsl 
django-elasticsearch-dsl
elasticsearch
python-logstash